Refactor regression detector into submodules
