+++ tools/copybook-bench/src/regression.rs
@@ -28,2093 +28,19 @@
     clippy::needless_borrow
 )]
 
-use sha2::{Digest, Sha256};
-use std::collections::HashMap;
-use std::fmt;
-use std::time::{Duration, Instant, SystemTime};
-
-/// Performance regression detector with CI integration
-pub struct PerformanceRegressionDetector {
-    baseline_repository: BaselineRepository,
-    statistical_analyzer: StatisticalRegressionAnalyzer,
-    ci_integrator: CiIntegrator,
-    alert_system: AlertSystem,
-}
-
-/// Repository for managing performance baselines
-#[derive(Debug, Clone)]
-pub struct BaselineRepository {
-    storage_backend: StorageBackend,
-    baseline_metadata: HashMap<String, BaselineMetadata>,
-    retention_policy: BaselineRetentionPolicy,
-}
-
-/// Storage backend for baselines
-#[derive(Debug, Clone)]
-pub enum StorageBackend {
-    FileSystem {
-        root_path: String,
-    },
-    Database {
-        connection_string: String,
-    },
-    CloudStorage {
-        bucket_name: String,
-        region: String,
-    },
-    GitRepository {
-        repository_url: String,
-        branch: String,
-    },
-}
-
-/// Baseline metadata with environment context
-#[derive(Debug, Clone)]
-pub struct BaselineMetadata {
-    pub baseline_id: String,
-    pub creation_timestamp: SystemTime,
-    pub git_commit_hash: Option<String>,
-    pub environment_info: EnvironmentInfo,
-    pub performance_metrics: PerformanceMetrics,
-    pub statistical_properties: StatisticalProperties,
-    pub validation_status: BaselineValidationStatus,
-}
-
-/// Environment information for baseline context
-#[derive(Debug, Clone)]
-pub struct EnvironmentInfo {
-    pub rust_version: String,
-    pub target_triple: String,
-    pub cpu_info: CpuInfo,
-    pub memory_info: MemoryInfo,
-    pub build_configuration: BuildConfiguration,
-}
-
-/// CPU information
-#[derive(Debug, Clone)]
-pub struct CpuInfo {
-    pub model: String,
-    pub cores: usize,
-    pub frequency_mhz: u32,
-    pub cache_size_kb: u32,
-}
-
-/// Memory information
-#[derive(Debug, Clone)]
-pub struct MemoryInfo {
-    pub total_gb: f64,
-    pub available_gb: f64,
-    pub memory_type: String,
-}
-
-/// Build configuration details
-#[derive(Debug, Clone)]
-pub struct BuildConfiguration {
-    pub optimization_level: OptimizationLevel,
-    pub debug_info: bool,
-    pub target_cpu: String,
-    pub features_enabled: Vec<String>,
-}
-
-/// Optimization levels
-#[derive(Debug, Clone)]
-pub enum OptimizationLevel {
-    Debug,
-    Release,
-    ReleaseWithDebugInfo,
-}
-
-/// Performance metrics for regression analysis
-#[derive(Debug, Clone)]
-pub struct PerformanceMetrics {
-    pub display_throughput: ThroughputMetrics,
-    pub comp3_throughput: ThroughputMetrics,
-    pub memory_usage: MemoryUsageMetrics,
-    pub latency_metrics: LatencyMetrics,
-}
-
-/// Throughput metrics
-#[derive(Debug, Clone)]
-pub struct ThroughputMetrics {
-    pub mean: f64,
-    pub median: f64,
-    pub percentile_95: f64,
-    pub percentile_99: f64,
-    pub std_deviation: f64,
-    pub min: f64,
-    pub max: f64,
-}
-
-/// Memory usage metrics
-#[derive(Debug, Clone)]
-pub struct MemoryUsageMetrics {
-    pub peak_memory_mb: f64,
-    pub average_memory_mb: f64,
-    pub steady_state_memory_mb: f64,
-    pub memory_variance: f64,
-}
-
-/// Latency metrics
-#[derive(Debug, Clone)]
-pub struct LatencyMetrics {
-    pub p50_ms: f64,
-    pub p95_ms: f64,
-    pub p99_ms: f64,
-    pub p999_ms: f64,
-    pub max_latency_ms: f64,
-}
-
-/// Statistical properties of metrics
-#[derive(Debug, Clone)]
-pub struct StatisticalProperties {
-    pub sample_size: usize,
-    pub confidence_interval_95: ConfidenceInterval,
-    pub statistical_significance: bool,
-    pub normality_test_passed: bool,
-}
-
-/// Confidence interval
-#[derive(Debug, Clone)]
-pub struct ConfidenceInterval {
-    pub lower_bound: f64,
-    pub upper_bound: f64,
-    pub confidence_level: f64,
-}
-
-/// Baseline validation status
-#[derive(Debug, Clone)]
-pub enum BaselineValidationStatus {
-    Valid,
-    UnderReview,
-    Invalid { reason: String },
-    Superseded { new_baseline_id: String },
-}
-
-/// Baseline retention policy
-#[derive(Debug, Clone)]
-pub struct BaselineRetentionPolicy {
-    pub max_baselines_per_branch: usize,
-    pub retention_days: u32,
-    pub archive_after_days: u32,
-}
-
-/// Statistical regression analyzer
-pub struct StatisticalRegressionAnalyzer {
-    variance_tolerance: f64,
-    significance_level: f64,
-    minimum_sample_size: usize,
-    outlier_detection: OutlierDetectionConfig,
-}
-
-/// Outlier detection configuration
-#[derive(Debug, Clone)]
-pub struct OutlierDetectionConfig {
-    pub method: OutlierDetectionMethod,
-    pub threshold: f64,
-    pub action: OutlierAction,
-}
-
-/// Outlier detection methods
-#[derive(Debug, Clone)]
-pub enum OutlierDetectionMethod {
-    IQR,
-    ZScore,
-    ModifiedZScore,
-    IsolationForest,
-}
-
-/// Actions for outliers
-#[derive(Debug, Clone)]
-pub enum OutlierAction {
-    Remove,
-    Flag,
-    Transform,
-    Ignore,
-}
-
-/// Regression analysis result
-#[derive(Debug)]
-pub struct RegressionAnalysis {
-    pub status: RegressionStatus,
-    pub metrics_comparison: MetricsComparison,
-    pub statistical_tests: StatisticalTestResults,
-    pub confidence_score: f64,
-    pub recommendations: Vec<RecommendationAction>,
-}
-
-/// Regression status
-#[derive(Debug)]
-pub enum RegressionStatus {
-    NoRegression,
-    MinorRegression { severity: RegressionSeverity },
-    MajorRegression { severity: RegressionSeverity },
-    CriticalRegression { severity: RegressionSeverity },
-    Improvement { magnitude: f64 },
-}
-
-/// Regression severity
-#[derive(Debug, Clone)]
-pub enum RegressionSeverity {
-    Low,
-    Medium,
-    High,
-    Critical,
-}
-
-/// Metrics comparison
-#[derive(Debug)]
-pub struct MetricsComparison {
-    pub throughput_changes: Vec<MetricChange>,
-    pub memory_changes: Vec<MetricChange>,
-    pub latency_changes: Vec<MetricChange>,
-    pub overall_change_percent: f64,
-}
-
-/// Individual metric change
-#[derive(Debug)]
-pub struct MetricChange {
-    pub metric_name: String,
-    pub baseline_value: f64,
-    pub current_value: f64,
-    pub change_percent: f64,
-    pub change_direction: ChangeDirection,
-    pub statistical_significance: bool,
-}
-
-/// Direction of change
-#[derive(Debug)]
-pub enum ChangeDirection {
-    Improvement,
-    Degradation,
-    Neutral,
-}
-
-/// Statistical test results
-#[derive(Debug)]
-pub struct StatisticalTestResults {
-    pub t_test_result: TTestResult,
-    pub mann_whitney_u_result: MannWhitneyResult,
-    pub effect_size: EffectSize,
-    pub power_analysis: PowerAnalysisResult,
-}
-
-/// T-test result
-#[derive(Debug)]
-pub struct TTestResult {
-    pub statistic: f64,
-    pub p_value: f64,
-    pub degrees_of_freedom: usize,
-    pub significant: bool,
-}
-
-/// Mann-Whitney U test result
-#[derive(Debug)]
-pub struct MannWhitneyResult {
-    pub u_statistic: f64,
-    pub p_value: f64,
-    pub significant: bool,
-}
-
-/// Effect size measures
-#[derive(Debug)]
-pub struct EffectSize {
-    pub cohens_d: f64,
-    pub glass_delta: f64,
-    pub hedges_g: f64,
-    pub interpretation: EffectSizeInterpretation,
-}
-
-/// Effect size interpretation
-#[derive(Debug)]
-pub enum EffectSizeInterpretation {
-    Negligible,
-    Small,
-    Medium,
-    Large,
-    VeryLarge,
-}
-
-/// Power analysis result
-#[derive(Debug)]
-pub struct PowerAnalysisResult {
-    pub power: f64,
-    pub required_sample_size: usize,
-    pub adequate_power: bool,
-}
-
-/// Recommendation actions
-#[derive(Debug)]
-pub enum RecommendationAction {
-    AcceptRegression { reason: String },
-    RejectChanges { reason: String },
-    InvestigateRegression { focus_areas: Vec<String> },
-    IncreaseBaseline { new_baseline_justification: String },
-    OptimizePerformance { suggested_areas: Vec<String> },
-}
-
-/// CI integrator for automated checks
-pub struct CiIntegrator {
-    performance_gates: Vec<PerformanceGate>,
-    notification_config: NotificationConfig,
-    integration_config: CiIntegrationConfig,
-}
-
-/// Performance gate configuration
-#[derive(Debug)]
-pub struct PerformanceGate {
-    pub gate_id: String,
-    pub metric_type: GateMetricType,
-    pub threshold: GateThreshold,
-    pub action: GateAction,
-}
-
-/// Gate metric types
-#[derive(Debug)]
-pub enum GateMetricType {
-    DisplayThroughput,
-    Comp3Throughput,
-    MemoryUsage,
-    Latency,
-    Overall,
-}
-
-/// Gate threshold
-#[derive(Debug)]
-pub struct GateThreshold {
-    pub max_regression_percent: f64,
-    pub confidence_level: f64,
-    pub require_statistical_significance: bool,
-}
-
-/// Gate actions
-#[derive(Debug)]
-pub enum GateAction {
-    Block,
-    Warn,
-    Notify,
-    Skip,
-}
-
-/// Notification configuration
-#[derive(Debug)]
-pub struct NotificationConfig {
-    pub slack_webhook: Option<String>,
-    pub email_recipients: Vec<String>,
-    pub github_integration: Option<GitHubIntegrationConfig>,
-}
-
-/// GitHub integration configuration
-#[derive(Debug)]
-pub struct GitHubIntegrationConfig {
-    pub token: String,
-    pub repository: String,
-    pub create_issues: bool,
-    pub pr_comments: bool,
-}
-
-/// CI integration configuration
-#[derive(Debug)]
-pub struct CiIntegrationConfig {
-    pub platforms: Vec<CiPlatform>,
-    pub artifact_storage: ArtifactStorageConfig,
-    pub reporting_config: ReportingConfig,
-}
-
-/// CI platforms
-#[derive(Debug)]
-pub enum CiPlatform {
-    GitHubActions,
-    GitLabCI,
-    JenkinsCI,
-    BuildKite,
-    TeamCity,
-}
-
-/// Artifact storage configuration
-#[derive(Debug)]
-pub struct ArtifactStorageConfig {
-    pub store_raw_data: bool,
-    pub store_analysis_reports: bool,
-    pub retention_days: u32,
-}
-
-/// Reporting configuration
-#[derive(Debug)]
-pub struct ReportingConfig {
-    pub generate_html_reports: bool,
-    pub generate_json_reports: bool,
-    pub include_trend_analysis: bool,
-}
-
-/// CI check result
-#[derive(Debug)]
-pub struct CiCheckResult {
-    pub overall_status: CiStatus,
-    pub gate_results: Vec<GateResult>,
-    pub analysis_summary: AnalysisSummary,
-    pub recommendations: Vec<RecommendationAction>,
-}
-
-/// CI status
-#[derive(Debug)]
-pub enum CiStatus {
-    Passed,
-    Failed { reason: String },
-    Warning { message: String },
-    Skipped { reason: String },
-}
-
-/// Gate result
-#[derive(Debug)]
-pub struct GateResult {
-    pub gate_id: String,
-    pub status: GateStatus,
-    pub measured_value: f64,
-    pub threshold_value: f64,
-    pub message: String,
-}
-
-/// Gate status
-#[derive(Debug)]
-pub enum GateStatus {
-    Passed,
-    Failed,
-    Warning,
-    Skipped,
-}
-
-/// Analysis summary
-#[derive(Debug)]
-pub struct AnalysisSummary {
-    pub total_metrics_analyzed: usize,
-    pub regressions_detected: usize,
-    pub improvements_detected: usize,
-    pub confidence_score: f64,
-    pub analysis_duration: Duration,
-}
-
-/// Alert system for notifications
-pub struct AlertSystem {
-    alert_policies: Vec<AlertPolicy>,
-    escalation_manager: EscalationManager,
-    notification_channels: Vec<NotificationChannel>,
-}
-
-/// Alert policy
-#[derive(Debug, Clone)]
-pub struct AlertPolicy {
-    pub policy_id: String,
-    pub trigger_conditions: Vec<AlertTrigger>,
-    pub severity: AlertSeverity,
-    pub escalation_policy_id: Option<String>,
-}
-
-/// Alert triggers
-#[derive(Debug, Clone)]
-pub enum AlertTrigger {
-    RegressionDetected { min_severity: RegressionSeverity },
-    ThresholdExceeded { metric: String, threshold: f64 },
-    RepeatedRegressions { count: u32 },
-    CriticalMetricFailure,
-}
-
-/// Alert severity
-#[derive(Debug, Clone)]
-pub enum AlertSeverity {
-    Info,
-    Warning,
-    Error,
-    Critical,
-}
-
-/// Escalation manager
-pub struct EscalationManager {
-    escalation_policies: HashMap<String, EscalationPolicy>,
-    active_escalations: HashMap<String, ActiveEscalation>,
-}
-
-/// Escalation policy
-#[derive(Debug)]
-pub struct EscalationPolicy {
-    pub policy_id: String,
-    pub escalation_steps: Vec<EscalationStep>,
-    pub max_escalation_level: u32,
-}
-
-/// Escalation step
-#[derive(Debug)]
-pub struct EscalationStep {
-    pub step_number: u32,
-    pub delay: Duration,
-    pub notification_targets: Vec<String>,
-    pub required_acknowledgment: bool,
-}
-
-/// Active escalation
-#[derive(Debug)]
-pub struct ActiveEscalation {
-    pub escalation_id: String,
-    pub policy_id: String,
-    pub current_step: u32,
-    pub start_time: SystemTime,
-    pub acknowledged: bool,
-}
-
-/// Notification channels
-#[derive(Debug)]
-pub enum NotificationChannel {
-    Email {
-        recipients: Vec<String>,
-    },
-    Slack {
-        webhook_url: String,
-        channel: String,
-    },
-    Teams {
-        webhook_url: String,
-    },
-    PagerDuty {
-        integration_key: String,
-    },
-    Custom {
-        name: String,
-        config: HashMap<String, String>,
-    },
-}
-
-/// Alert result
-#[derive(Debug)]
-pub struct AlertResult {
-    pub alert_id: String,
-    pub policy_id: String,
-    pub severity: AlertSeverity,
-    pub message: String,
-    pub notifications_sent: Vec<NotificationResult>,
-    pub escalation_triggered: bool,
-}
-
-/// Notification result
-#[derive(Debug)]
-pub struct NotificationResult {
-    pub channel: String,
-    pub success: bool,
-    pub error_message: Option<String>,
-    pub delivery_time: Duration,
-}
-
-impl PerformanceRegressionDetector {
-    /// Create new performance regression detector
-    pub fn new() -> Self {
-        Self {
-            baseline_repository: BaselineRepository::new(),
-            statistical_analyzer: StatisticalRegressionAnalyzer::new(),
-            ci_integrator: CiIntegrator::new(),
-            alert_system: AlertSystem::new(),
-        }
-    }
-
-    /// Detect performance regression with comprehensive analysis
-    pub fn detect_performance_regression(
-        &mut self,
-        current_metrics: PerformanceMetrics,
-    ) -> Result<RegressionAnalysis, Box<dyn std::error::Error>> {
-        // Find compatible baseline for comparison
-        let environment = self.get_current_environment()?;
-        let baseline_metadata = self
-            .baseline_repository
-            .find_compatible_baseline(&environment)?
-            .ok_or("No compatible baseline found")?;
-
-        // Perform statistical analysis
-        let comparison = self
-            .statistical_analyzer
-            .compare_metrics(&baseline_metadata.performance_metrics, &current_metrics)?;
-
-        // Run statistical tests
-        let statistical_tests = self
-            .statistical_analyzer
-            .run_statistical_tests(&baseline_metadata.performance_metrics, &current_metrics)?;
-
-        // Determine regression status
-        let status = self.determine_regression_status(&comparison, &statistical_tests)?;
-
-        // Calculate confidence score
-        let confidence_score = self.calculate_confidence_score(&statistical_tests);
-
-        // Generate recommendations
-        let recommendations = self.generate_recommendations(&status, &comparison)?;
-
-        Ok(RegressionAnalysis {
-            status,
-            metrics_comparison: comparison,
-            statistical_tests,
-            confidence_score,
-            recommendations,
-        })
-    }
-
-    /// Establish new performance baseline
-    pub fn establish_baseline(
-        &mut self,
-        metrics: PerformanceMetrics,
-        environment: EnvironmentInfo,
-    ) -> Result<String, Box<dyn std::error::Error>> {
-        // Generate baseline ID
-        let baseline_id = self.generate_baseline_id(&environment);
-
-        // Calculate statistical properties
-        let statistical_properties = self
-            .statistical_analyzer
-            .calculate_statistical_properties(&metrics)?;
-
-        // Create baseline metadata
-        let baseline_metadata = BaselineMetadata {
-            baseline_id: baseline_id.clone(),
-            creation_timestamp: SystemTime::now(),
-            git_commit_hash: self.get_current_git_commit()?,
-            environment_info: environment,
-            performance_metrics: metrics,
-            statistical_properties,
-            validation_status: BaselineValidationStatus::Valid,
-        };
-
-        // Store baseline
-        self.baseline_repository.store_baseline(baseline_metadata)?;
-
-        Ok(baseline_id)
-    }
-
-    /// Execute CI performance check
-    pub fn execute_ci_performance_check(
-        &mut self,
-    ) -> Result<CiCheckResult, Box<dyn std::error::Error>> {
-        let start_time = Instant::now();
-
-        // Run performance measurements
-        let current_metrics = self.measure_current_performance()?;
-
-        // Detect regressions
-        let regression_analysis = self.detect_performance_regression(current_metrics)?;
-
-        // Execute performance gates
-        let gate_results = self
-            .ci_integrator
-            .execute_performance_gates(&regression_analysis)?;
-
-        // Determine overall status
-        let overall_status = self.determine_ci_status(&gate_results);
-
-        // Generate recommendations
-        let recommendations =
-            self.generate_ci_recommendations(&regression_analysis, &gate_results)?;
-
-        // Create analysis summary
-        let analysis_summary = AnalysisSummary {
-            total_metrics_analyzed: self.count_analyzed_metrics(&regression_analysis),
-            regressions_detected: self.count_regressions(&regression_analysis),
-            improvements_detected: self.count_improvements(&regression_analysis),
-            confidence_score: regression_analysis.confidence_score,
-            analysis_duration: start_time.elapsed(),
-        };
-
-        Ok(CiCheckResult {
-            overall_status,
-            gate_results,
-            analysis_summary,
-            recommendations,
-        })
-    }
-
-    /// Trigger performance alerts
-    pub fn trigger_alerts(
-        &mut self,
-        analysis: &RegressionAnalysis,
-    ) -> Result<Vec<AlertResult>, Box<dyn std::error::Error>> {
-        self.alert_system.evaluate_alerts(analysis)
-    }
-
-    // Helper methods
-
-    fn get_current_environment(&self) -> Result<EnvironmentInfo, Box<dyn std::error::Error>> {
-        Ok(EnvironmentInfo {
-            rust_version: std::env::var("RUSTC_VERSION").unwrap_or_else(|_| "unknown".to_string()),
-            target_triple: std::env::var("TARGET").unwrap_or_else(|_| "unknown".to_string()),
-            cpu_info: CpuInfo {
-                model: "Generic CPU".to_string(),
-                cores: num_cpus::get(),
-                frequency_mhz: 2400, // Default assumption
-                cache_size_kb: 8192, // Default assumption
-            },
-            memory_info: MemoryInfo {
-                total_gb: 16.0,    // Default assumption
-                available_gb: 8.0, // Default assumption
-                memory_type: "DDR4".to_string(),
-            },
-            build_configuration: BuildConfiguration {
-                optimization_level: if cfg!(debug_assertions) {
-                    OptimizationLevel::Debug
-                } else {
-                    OptimizationLevel::Release
-                },
-                debug_info: cfg!(debug_assertions),
-                target_cpu: std::env::var("TARGET_CPU").unwrap_or_else(|_| "native".to_string()),
-                features_enabled: Vec::new(),
-            },
-        })
-    }
-
-    fn generate_baseline_id(&self, environment: &EnvironmentInfo) -> String {
-        let mut hasher = Sha256::new();
-        hasher.update(&environment.rust_version);
-        hasher.update(&environment.target_triple);
-        hasher.update(format!(
-            "{}",
-            SystemTime::now()
-                .duration_since(SystemTime::UNIX_EPOCH)
-                .unwrap_or_else(|_| std::time::Duration::from_secs(0))
-                .as_secs()
-        ));
-        format!("baseline_{:x}", hasher.finalize())[..16].to_string()
-    }
-
-    fn get_current_git_commit(&self) -> Result<Option<String>, Box<dyn std::error::Error>> {
-        // Simplified - real implementation would use git2 or similar
-        Ok(std::env::var("GIT_COMMIT").ok())
-    }
-
-    fn determine_regression_status(
-        &self,
-        comparison: &MetricsComparison,
-        tests: &StatisticalTestResults,
-    ) -> Result<RegressionStatus, Box<dyn std::error::Error>> {
-        let max_degradation = comparison
-            .throughput_changes
-            .iter()
-            .chain(comparison.memory_changes.iter())
-            .chain(comparison.latency_changes.iter())
-            .filter(|change| matches!(change.change_direction, ChangeDirection::Degradation))
-            .map(|change| change.change_percent)
-            .fold(0.0f64, f64::max);
-        let max_improvement = comparison
-            .throughput_changes
-            .iter()
-            .chain(comparison.memory_changes.iter())
-            .chain(comparison.latency_changes.iter())
-            .filter(|change| matches!(change.change_direction, ChangeDirection::Improvement))
-            .map(|change| change.change_percent)
-            .fold(0.0f64, f64::max);
-
-        let is_significant = tests.t_test_result.significant;
-
-        if !is_significant {
-            Ok(RegressionStatus::NoRegression)
-        } else if max_improvement >= 2.0 && max_degradation < 2.0 {
-            Ok(RegressionStatus::Improvement {
-                magnitude: max_improvement,
-            })
-        } else if max_degradation < 2.0 {
-            Ok(RegressionStatus::NoRegression)
-        } else if max_degradation < 5.0 {
-            Ok(RegressionStatus::MinorRegression {
-                severity: RegressionSeverity::Low,
-            })
-        } else if max_degradation < 15.0 {
-            Ok(RegressionStatus::MajorRegression {
-                severity: RegressionSeverity::Medium,
-            })
-        } else {
-            Ok(RegressionStatus::CriticalRegression {
-                severity: RegressionSeverity::High,
-            })
-        }
-    }
-
-    fn calculate_confidence_score(&self, tests: &StatisticalTestResults) -> f64 {
-        let mut score = 0.0;
-
-        // T-test contribution
-        if tests.t_test_result.significant {
-            score += 0.4 * (1.0 - tests.t_test_result.p_value);
-        }
-
-        // Mann-Whitney U contribution
-        if tests.mann_whitney_u_result.significant {
-            score += 0.3 * (1.0 - tests.mann_whitney_u_result.p_value);
-        }
-
-        // Effect size contribution
-        score += 0.2
-            * match tests.effect_size.interpretation {
-                EffectSizeInterpretation::Negligible => 0.1,
-                EffectSizeInterpretation::Small => 0.3,
-                EffectSizeInterpretation::Medium => 0.6,
-                EffectSizeInterpretation::Large => 0.8,
-                EffectSizeInterpretation::VeryLarge => 1.0,
-            };
-
-        // Power analysis contribution
-        if tests.power_analysis.adequate_power {
-            score += 0.1;
-        }
-
-        score.min(1.0)
-    }
-
-    fn generate_recommendations(
-        &self,
-        status: &RegressionStatus,
-        _comparison: &MetricsComparison,
-    ) -> Result<Vec<RecommendationAction>, Box<dyn std::error::Error>> {
-        let mut recommendations = Vec::new();
-
-        match status {
-            RegressionStatus::NoRegression => {
-                recommendations.push(RecommendationAction::AcceptRegression {
-                    reason: "Performance metrics within acceptable variance".to_string(),
-                });
-            }
-            RegressionStatus::MinorRegression { .. } => {
-                recommendations.push(RecommendationAction::InvestigateRegression {
-                    focus_areas: vec![
-                        "Algorithm efficiency".to_string(),
-                        "Memory allocation patterns".to_string(),
-                    ],
-                });
-            }
-            RegressionStatus::MajorRegression { .. } => {
-                recommendations.push(RecommendationAction::RejectChanges {
-                    reason: "Major performance regression detected".to_string(),
-                });
-                recommendations.push(RecommendationAction::OptimizePerformance {
-                    suggested_areas: vec![
-                        "Hot paths optimization".to_string(),
-                        "Data structure efficiency".to_string(),
-                    ],
-                });
-            }
-            RegressionStatus::CriticalRegression { .. } => {
-                recommendations.push(RecommendationAction::RejectChanges {
-                    reason: "Critical performance regression - immediate action required"
-                        .to_string(),
-                });
-            }
-            RegressionStatus::Improvement { .. } => {
-                recommendations.push(RecommendationAction::AcceptRegression {
-                    reason: "Performance improvement detected".to_string(),
-                });
-            }
-        }
-
-        Ok(recommendations)
-    }
-
-    fn measure_current_performance(
-        &self,
-    ) -> Result<PerformanceMetrics, Box<dyn std::error::Error>> {
-        // Simplified implementation - real version would run actual benchmarks
-        Ok(PerformanceMetrics {
-            display_throughput: ThroughputMetrics {
-                mean: 4200.0,
-                median: 4150.0,
-                percentile_95: 4800.0,
-                percentile_99: 5200.0,
-                std_deviation: 250.0,
-                min: 3800.0,
-                max: 5500.0,
-            },
-            comp3_throughput: ThroughputMetrics {
-                mean: 580.0,
-                median: 575.0,
-                percentile_95: 620.0,
-                percentile_99: 650.0,
-                std_deviation: 25.0,
-                min: 540.0,
-                max: 680.0,
-            },
-            memory_usage: MemoryUsageMetrics {
-                peak_memory_mb: 245.0,
-                average_memory_mb: 180.0,
-                steady_state_memory_mb: 165.0,
-                memory_variance: 12.5,
-            },
-            latency_metrics: LatencyMetrics {
-                p50_ms: 0.24,
-                p95_ms: 0.45,
-                p99_ms: 0.68,
-                p999_ms: 1.2,
-                max_latency_ms: 2.1,
-            },
-        })
-    }
-
-    fn determine_ci_status(&self, gate_results: &[GateResult]) -> CiStatus {
-        let failed_count = gate_results
-            .iter()
-            .filter(|r| matches!(r.status, GateStatus::Failed))
-            .count();
-        let warning_count = gate_results
-            .iter()
-            .filter(|r| matches!(r.status, GateStatus::Warning))
-            .count();
-
-        if failed_count > 0 {
-            CiStatus::Failed {
-                reason: format!("{} performance gate(s) failed", failed_count),
-            }
-        } else if warning_count > 0 {
-            CiStatus::Warning {
-                message: format!("{} performance gate(s) have warnings", warning_count),
-            }
-        } else {
-            CiStatus::Passed
-        }
-    }
-
-    fn generate_ci_recommendations(
-        &self,
-        _analysis: &RegressionAnalysis,
-        _gates: &[GateResult],
-    ) -> Result<Vec<RecommendationAction>, Box<dyn std::error::Error>> {
-        // Simplified implementation
-        Ok(vec![RecommendationAction::InvestigateRegression {
-            focus_areas: vec!["CI environment consistency".to_string()],
-        }])
-    }
-
-    fn count_analyzed_metrics(&self, analysis: &RegressionAnalysis) -> usize {
-        analysis.metrics_comparison.throughput_changes.len()
-            + analysis.metrics_comparison.memory_changes.len()
-            + analysis.metrics_comparison.latency_changes.len()
-    }
-
-    fn count_regressions(&self, analysis: &RegressionAnalysis) -> usize {
-        match analysis.status {
-            RegressionStatus::NoRegression | RegressionStatus::Improvement { .. } => 0,
-            _ => 1,
-        }
-    }
-
-    fn count_improvements(&self, analysis: &RegressionAnalysis) -> usize {
-        match analysis.status {
-            RegressionStatus::Improvement { .. } => 1,
-            _ => 0,
-        }
-    }
-}
-
-impl Default for PerformanceRegressionDetector {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl BaselineRepository {
-    pub fn new() -> Self {
-        Self {
-            storage_backend: StorageBackend::FileSystem {
-                root_path: std::env::var("BASELINE_STORAGE_PATH")
-                    .unwrap_or_else(|_| "/tmp/copybook-baselines".to_string()),
-            },
-            baseline_metadata: HashMap::new(),
-            retention_policy: BaselineRetentionPolicy {
-                max_baselines_per_branch: 10,
-                retention_days: 90,
-                archive_after_days: 30,
-            },
-        }
-    }
-
-    pub fn store_baseline(
-        &mut self,
-        baseline: BaselineMetadata,
-    ) -> Result<String, Box<dyn std::error::Error>> {
-        let baseline_id = baseline.baseline_id.clone();
-        self.baseline_metadata.insert(baseline_id.clone(), baseline);
-        Ok(baseline_id)
-    }
-
-    pub fn load_baseline(
-        &self,
-        baseline_id: &str,
-    ) -> Result<BaselineMetadata, Box<dyn std::error::Error>> {
-        self.baseline_metadata
-            .get(baseline_id)
-            .cloned()
-            .ok_or_else(|| format!("Baseline {} not found", baseline_id).into())
-    }
-
-    pub fn find_compatible_baseline(
-        &self,
-        environment: &EnvironmentInfo,
-    ) -> Result<Option<BaselineMetadata>, Box<dyn std::error::Error>> {
-        // Find baseline with matching environment characteristics
-        for baseline in self.baseline_metadata.values() {
-            if self.is_environment_compatible(&baseline.environment_info, environment) {
-                return Ok(Some(baseline.clone()));
-            }
-        }
-        Ok(None)
-    }
-
-    fn is_environment_compatible(
-        &self,
-        baseline_env: &EnvironmentInfo,
-        current_env: &EnvironmentInfo,
-    ) -> bool {
-        // Simplified compatibility check
-        baseline_env.target_triple == current_env.target_triple
-            && matches!(
-                (
-                    &baseline_env.build_configuration.optimization_level,
-                    &current_env.build_configuration.optimization_level
-                ),
-                (OptimizationLevel::Release, OptimizationLevel::Release)
-                    | (OptimizationLevel::Debug, OptimizationLevel::Debug)
-            )
-    }
-}
-
-impl StatisticalRegressionAnalyzer {
-    pub fn new() -> Self {
-        Self {
-            variance_tolerance: 0.02, // 2% tolerance
-            significance_level: 0.05,
-            minimum_sample_size: 30,
-            outlier_detection: OutlierDetectionConfig {
-                method: OutlierDetectionMethod::IQR,
-                threshold: 1.5,
-                action: OutlierAction::Flag,
-            },
-        }
-    }
-
-    pub fn compare_metrics(
-        &self,
-        baseline: &PerformanceMetrics,
-        current: &PerformanceMetrics,
-    ) -> Result<MetricsComparison, Box<dyn std::error::Error>> {
-        let mut throughput_changes = Vec::new();
-        let mut memory_changes = Vec::new();
-        let mut latency_changes = Vec::new();
-
-        // Compare display throughput
-        throughput_changes.push(self.calculate_metric_change(
-            "display_throughput_mean",
-            baseline.display_throughput.mean,
-            current.display_throughput.mean,
-        ));
-
-        // Compare COMP-3 throughput
-        throughput_changes.push(self.calculate_metric_change(
-            "comp3_throughput_mean",
-            baseline.comp3_throughput.mean,
-            current.comp3_throughput.mean,
-        ));
-
-        // Compare memory usage
-        memory_changes.push(self.calculate_metric_change(
-            "peak_memory_mb",
-            baseline.memory_usage.peak_memory_mb,
-            current.memory_usage.peak_memory_mb,
-        ));
-
-        // Compare latency
-        latency_changes.push(self.calculate_metric_change(
-            "p95_latency_ms",
-            baseline.latency_metrics.p95_ms,
-            current.latency_metrics.p95_ms,
-        ));
-
-        let overall_change_percent =
-            self.calculate_overall_change(&throughput_changes, &memory_changes, &latency_changes);
-
-        Ok(MetricsComparison {
-            throughput_changes,
-            memory_changes,
-            latency_changes,
-            overall_change_percent,
-        })
-    }
-
-    pub fn run_statistical_tests(
-        &self,
-        baseline: &PerformanceMetrics,
-        current: &PerformanceMetrics,
-    ) -> Result<StatisticalTestResults, Box<dyn std::error::Error>> {
-        // Simplified statistical tests - real implementation would use proper statistical libraries
-        let t_test_result = self.perform_t_test(
-            baseline.display_throughput.mean,
-            current.display_throughput.mean,
-        )?;
-        let mann_whitney_u_result = self.perform_mann_whitney_u_test(baseline, current)?;
-        let effect_size = self.calculate_effect_size(baseline, current)?;
-        let power_analysis = self.perform_power_analysis()?;
-
-        Ok(StatisticalTestResults {
-            t_test_result,
-            mann_whitney_u_result,
-            effect_size,
-            power_analysis,
-        })
-    }
-
-    pub fn calculate_statistical_properties(
-        &self,
-        _metrics: &PerformanceMetrics,
-    ) -> Result<StatisticalProperties, Box<dyn std::error::Error>> {
-        Ok(StatisticalProperties {
-            sample_size: 100, // Assumed
-            confidence_interval_95: ConfidenceInterval {
-                lower_bound: 4000.0,
-                upper_bound: 4400.0,
-                confidence_level: 0.95,
-            },
-            statistical_significance: true,
-            normality_test_passed: true,
-        })
-    }
-
-    fn calculate_metric_change(
-        &self,
-        name: &str,
-        baseline_value: f64,
-        current_value: f64,
-    ) -> MetricChange {
-        let change_percent = ((current_value - baseline_value) / baseline_value) * 100.0;
-        let change_direction = if change_percent > self.variance_tolerance * 100.0 {
-            if name.contains("throughput") {
-                ChangeDirection::Improvement
-            } else {
-                ChangeDirection::Degradation
-            }
-        } else if change_percent < -self.variance_tolerance * 100.0 {
-            if name.contains("throughput") {
-                ChangeDirection::Degradation
-            } else {
-                ChangeDirection::Improvement
-            }
-        } else {
-            ChangeDirection::Neutral
-        };
-
-        MetricChange {
-            metric_name: name.to_string(),
-            baseline_value,
-            current_value,
-            change_percent: change_percent.abs(),
-            change_direction,
-            statistical_significance: change_percent.abs() > self.variance_tolerance * 100.0,
-        }
-    }
-
-    fn calculate_overall_change(
-        &self,
-        throughput: &[MetricChange],
-        memory: &[MetricChange],
-        latency: &[MetricChange],
-    ) -> f64 {
-        let all_changes: Vec<f64> = throughput
-            .iter()
-            .chain(memory.iter())
-            .chain(latency.iter())
-            .map(|change| change.change_percent)
-            .collect();
-
-        if all_changes.is_empty() {
-            0.0
-        } else {
-            all_changes.iter().sum::<f64>() / all_changes.len() as f64
-        }
-    }
-
-    fn perform_t_test(
-        &self,
-        baseline_mean: f64,
-        current_mean: f64,
-    ) -> Result<TTestResult, Box<dyn std::error::Error>> {
-        // Simplified t-test calculation
-        let pooled_variance = 250.0; // Assumed
-        let n1 = 50; // Assumed sample size
-        let n2 = 50; // Assumed sample size
-
-        let standard_error = (pooled_variance * (1.0 / n1 as f64 + 1.0 / n2 as f64)).sqrt();
-        let t_statistic = (baseline_mean - current_mean) / standard_error;
-        let degrees_of_freedom = n1 + n2 - 2;
-
-        // Simplified p-value calculation (would use proper statistical library)
-        let p_value = if t_statistic.abs() > 2.0 { 0.01 } else { 0.1 };
-        let significant = p_value < self.significance_level;
-
-        Ok(TTestResult {
-            statistic: t_statistic,
-            p_value,
-            degrees_of_freedom,
-            significant,
-        })
-    }
-
-    fn perform_mann_whitney_u_test(
-        &self,
-        _baseline: &PerformanceMetrics,
-        _current: &PerformanceMetrics,
-    ) -> Result<MannWhitneyResult, Box<dyn std::error::Error>> {
-        // Simplified Mann-Whitney U test
-        Ok(MannWhitneyResult {
-            u_statistic: 1200.0,
-            p_value: 0.05,
-            significant: true,
-        })
-    }
-
-    fn calculate_effect_size(
-        &self,
-        baseline: &PerformanceMetrics,
-        current: &PerformanceMetrics,
-    ) -> Result<EffectSize, Box<dyn std::error::Error>> {
-        let mean_diff = baseline.display_throughput.mean - current.display_throughput.mean;
-        let pooled_std = ((baseline.display_throughput.std_deviation.powi(2)
-            + current.display_throughput.std_deviation.powi(2))
-            / 2.0)
-            .sqrt();
-
-        let cohens_d = mean_diff / pooled_std;
-        let glass_delta = mean_diff / baseline.display_throughput.std_deviation;
-        let hedges_g = cohens_d * (1.0 - 3.0 / (4.0 * 98.0 - 9.0)); // Bias correction
-
-        let interpretation = match cohens_d.abs() {
-            d if d < 0.2 => EffectSizeInterpretation::Negligible,
-            d if d < 0.5 => EffectSizeInterpretation::Small,
-            d if d < 0.8 => EffectSizeInterpretation::Medium,
-            d if d < 1.2 => EffectSizeInterpretation::Large,
-            _ => EffectSizeInterpretation::VeryLarge,
-        };
-
-        Ok(EffectSize {
-            cohens_d,
-            glass_delta,
-            hedges_g,
-            interpretation,
-        })
-    }
-
-    fn perform_power_analysis(&self) -> Result<PowerAnalysisResult, Box<dyn std::error::Error>> {
-        Ok(PowerAnalysisResult {
-            power: 0.85,
-            required_sample_size: 64,
-            adequate_power: true,
-        })
-    }
-}
-
-impl CiIntegrator {
-    pub fn new() -> Self {
-        Self {
-            performance_gates: vec![
-                PerformanceGate {
-                    gate_id: "display_throughput".to_string(),
-                    metric_type: GateMetricType::DisplayThroughput,
-                    threshold: GateThreshold {
-                        max_regression_percent: 2.0,
-                        confidence_level: 0.95,
-                        require_statistical_significance: true,
-                    },
-                    action: GateAction::Block,
-                },
-                PerformanceGate {
-                    gate_id: "comp3_throughput".to_string(),
-                    metric_type: GateMetricType::Comp3Throughput,
-                    threshold: GateThreshold {
-                        max_regression_percent: 2.0,
-                        confidence_level: 0.95,
-                        require_statistical_significance: true,
-                    },
-                    action: GateAction::Block,
-                },
-            ],
-            notification_config: NotificationConfig {
-                slack_webhook: std::env::var("SLACK_WEBHOOK").ok(),
-                email_recipients: Vec::new(),
-                github_integration: None,
-            },
-            integration_config: CiIntegrationConfig {
-                platforms: vec![CiPlatform::GitHubActions],
-                artifact_storage: ArtifactStorageConfig {
-                    store_raw_data: true,
-                    store_analysis_reports: true,
-                    retention_days: 30,
-                },
-                reporting_config: ReportingConfig {
-                    generate_html_reports: true,
-                    generate_json_reports: true,
-                    include_trend_analysis: true,
-                },
-            },
-        }
-    }
-
-    pub fn execute_performance_gates(
-        &self,
-        analysis: &RegressionAnalysis,
-    ) -> Result<Vec<GateResult>, Box<dyn std::error::Error>> {
-        let mut results = Vec::new();
-
-        for gate in &self.performance_gates {
-            let result = self.evaluate_gate(gate, analysis)?;
-            results.push(result);
-        }
-
-        Ok(results)
-    }
-
-    fn evaluate_gate(
-        &self,
-        gate: &PerformanceGate,
-        analysis: &RegressionAnalysis,
-    ) -> Result<GateResult, Box<dyn std::error::Error>> {
-        let (measured_value, threshold_value) = match gate.metric_type {
-            GateMetricType::DisplayThroughput => {
-                let change = analysis
-                    .metrics_comparison
-                    .throughput_changes
-                    .iter()
-                    .find(|c| c.metric_name.contains("display"))
-                    .map_or(0.0, |c| c.change_percent);
-                (change, gate.threshold.max_regression_percent)
-            }
-            GateMetricType::Comp3Throughput => {
-                let change = analysis
-                    .metrics_comparison
-                    .throughput_changes
-                    .iter()
-                    .find(|c| c.metric_name.contains("comp3"))
-                    .map_or(0.0, |c| c.change_percent);
-                (change, gate.threshold.max_regression_percent)
-            }
-            _ => (0.0, gate.threshold.max_regression_percent), // Simplified
-        };
-
-        let status = if measured_value <= threshold_value {
-            GateStatus::Passed
-        } else if measured_value <= threshold_value * 1.5 {
-            GateStatus::Warning
-        } else {
-            GateStatus::Failed
-        };
-
-        let message = format!(
-            "Gate {}: measured {:.2}%, threshold {:.2}%",
-            gate.gate_id, measured_value, threshold_value
-        );
-
-        Ok(GateResult {
-            gate_id: gate.gate_id.clone(),
-            status,
-            measured_value,
-            threshold_value,
-            message,
-        })
-    }
-}
-
-impl AlertSystem {
-    pub fn new() -> Self {
-        Self {
-            alert_policies: vec![AlertPolicy {
-                policy_id: "regression_alert".to_string(),
-                trigger_conditions: vec![AlertTrigger::RegressionDetected {
-                    min_severity: RegressionSeverity::Medium,
-                }],
-                severity: AlertSeverity::Warning,
-                escalation_policy_id: Some("standard_escalation".to_string()),
-            }],
-            escalation_manager: EscalationManager::new(),
-            notification_channels: vec![NotificationChannel::Slack {
-                webhook_url: std::env::var("SLACK_WEBHOOK").unwrap_or_default(),
-                channel: "#performance".to_string(),
-            }],
-        }
-    }
-
-    pub fn evaluate_alerts(
-        &mut self,
-        analysis: &RegressionAnalysis,
-    ) -> Result<Vec<AlertResult>, Box<dyn std::error::Error>> {
-        let mut results = Vec::new();
-
-        // Clone policies to avoid borrow checker issues
-        let policies = self.alert_policies.clone();
-        for policy in &policies {
-            if self.should_trigger_alert(&policy, analysis) {
-                let alert_result = self.trigger_alert(&policy, analysis)?;
-                results.push(alert_result);
-            }
-        }
-
-        Ok(results)
-    }
-
-    fn should_trigger_alert(&self, policy: &AlertPolicy, analysis: &RegressionAnalysis) -> bool {
-        for trigger in &policy.trigger_conditions {
-            match trigger {
-                AlertTrigger::RegressionDetected { min_severity } => {
-                    if let RegressionStatus::MajorRegression { severity }
-                    | RegressionStatus::CriticalRegression { severity } = &analysis.status
-                    {
-                        return matches!(
-                            (min_severity, severity),
-                            (RegressionSeverity::Low, _)
-                                | (
-                                    RegressionSeverity::Medium,
-                                    RegressionSeverity::Medium
-                                        | RegressionSeverity::High
-                                        | RegressionSeverity::Critical
-                                )
-                                | (
-                                    RegressionSeverity::High,
-                                    RegressionSeverity::High | RegressionSeverity::Critical
-                                )
-                                | (RegressionSeverity::Critical, RegressionSeverity::Critical)
-                        );
-                    }
-                }
-                _ => {} // Other triggers not implemented in this simplified version
-            }
-        }
-        false
-    }
-
-    fn trigger_alert(
-        &mut self,
-        policy: &AlertPolicy,
-        _analysis: &RegressionAnalysis,
-    ) -> Result<AlertResult, Box<dyn std::error::Error>> {
-        let alert_id = format!(
-            "alert_{}",
-            SystemTime::now()
-                .duration_since(SystemTime::UNIX_EPOCH)?
-                .as_secs()
-        );
-
-        let notifications_sent = self.send_notifications(&alert_id, policy)?;
-
-        Ok(AlertResult {
-            alert_id,
-            policy_id: policy.policy_id.clone(),
-            severity: policy.severity.clone(),
-            message: "Performance regression detected".to_string(),
-            notifications_sent,
-            escalation_triggered: false,
-        })
-    }
-
-    fn send_notifications(
-        &self,
-        _alert_id: &str,
-        _policy: &AlertPolicy,
-    ) -> Result<Vec<NotificationResult>, Box<dyn std::error::Error>> {
-        // Simplified notification sending
-        Ok(vec![NotificationResult {
-            channel: "slack".to_string(),
-            success: true,
-            error_message: None,
-            delivery_time: Duration::from_millis(150),
-        }])
-    }
-}
-
-impl EscalationManager {
-    pub fn new() -> Self {
-        Self {
-            escalation_policies: HashMap::new(),
-            active_escalations: HashMap::new(),
-        }
-    }
-}
-
-/// Error types for regression detection
-#[derive(Debug)]
-pub enum RegressionDetectionError {
-    BaselineNotFound(String),
-    InsufficientData(String),
-    StatisticalAnalysisError(String),
-    StorageError(String),
-    EnvironmentError(String),
-}
-
-impl fmt::Display for RegressionDetectionError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::BaselineNotFound(msg) => write!(f, "Baseline not found: {msg}"),
-            Self::InsufficientData(msg) => write!(f, "Insufficient data: {msg}"),
-            Self::StatisticalAnalysisError(msg) => write!(f, "Statistical analysis error: {msg}"),
-            Self::StorageError(msg) => write!(f, "Storage error: {msg}"),
-            Self::EnvironmentError(msg) => write!(f, "Environment error: {msg}"),
-        }
-    }
-}
-
-impl std::error::Error for RegressionDetectionError {}
-
-/// Utility functions for performance regression detection
-pub mod utils {
-    use super::*;
-
-    /// Create standard regression detector with default configuration
-    pub fn create_standard_detector() -> PerformanceRegressionDetector {
-        PerformanceRegressionDetector::new()
-    }
-
-    /// Create detector for CI environment with strict thresholds
-    pub fn create_ci_detector() -> PerformanceRegressionDetector {
-        let mut detector = PerformanceRegressionDetector::new();
-        // Configure stricter thresholds for CI
-        detector.statistical_analyzer.variance_tolerance = 0.015; // 1.5%
-        detector
-    }
-
-    /// Create detector for development environment with relaxed thresholds
-    pub fn create_dev_detector() -> PerformanceRegressionDetector {
-        let mut detector = PerformanceRegressionDetector::new();
-        // Configure more relaxed thresholds for development
-        detector.statistical_analyzer.variance_tolerance = 0.05; // 5%
-        detector
-    }
-}
-
-// Add dependency for CPU core detection
-#[cfg(test)]
-#[allow(clippy::expect_used)]
-#[allow(clippy::unwrap_used)]
-mod num_cpus {
-    pub fn get() -> usize {
-        4 // Default for testing
-    }
-}
-
-#[cfg(not(test))]
-mod num_cpus {
-    pub fn get() -> usize {
-        std::thread::available_parallelism().map_or(4, |n| n.get())
-    }
-}
+mod alerts;
+mod analyzer;
+mod ci;
+mod detector;
+mod error;
+mod repository;
+mod types;
+pub mod utils;
+
+pub use error::*;
+pub use types::*;
 
 #[cfg(test)]
 #[allow(clippy::expect_used)]
 #[allow(clippy::unwrap_used)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_detector_creation() {
-        let detector = PerformanceRegressionDetector::new();
-        assert!((detector.statistical_analyzer.variance_tolerance - 0.02).abs() < f64::EPSILON);
-    }
-
-    #[test]
-    fn test_baseline_repository() {
-        let mut repo = BaselineRepository::new();
-        let metrics = PerformanceMetrics {
-            display_throughput: ThroughputMetrics {
-                mean: 4200.0,
-                median: 4150.0,
-                percentile_95: 4800.0,
-                percentile_99: 5200.0,
-                std_deviation: 250.0,
-                min: 3800.0,
-                max: 5500.0,
-            },
-            comp3_throughput: ThroughputMetrics {
-                mean: 580.0,
-                median: 575.0,
-                percentile_95: 620.0,
-                percentile_99: 650.0,
-                std_deviation: 25.0,
-                min: 540.0,
-                max: 680.0,
-            },
-            memory_usage: MemoryUsageMetrics {
-                peak_memory_mb: 245.0,
-                average_memory_mb: 180.0,
-                steady_state_memory_mb: 165.0,
-                memory_variance: 12.5,
-            },
-            latency_metrics: LatencyMetrics {
-                p50_ms: 0.24,
-                p95_ms: 0.45,
-                p99_ms: 0.68,
-                p999_ms: 1.2,
-                max_latency_ms: 2.1,
-            },
-        };
-
-        let baseline = BaselineMetadata {
-            baseline_id: "test_baseline".to_string(),
-            creation_timestamp: SystemTime::now(),
-            git_commit_hash: None,
-            environment_info: EnvironmentInfo {
-                rust_version: "1.70.0".to_string(),
-                target_triple: "x86_64-unknown-linux-gnu".to_string(),
-                cpu_info: CpuInfo {
-                    model: "Test CPU".to_string(),
-                    cores: 4,
-                    frequency_mhz: 2400,
-                    cache_size_kb: 8192,
-                },
-                memory_info: MemoryInfo {
-                    total_gb: 16.0,
-                    available_gb: 8.0,
-                    memory_type: "DDR4".to_string(),
-                },
-                build_configuration: BuildConfiguration {
-                    optimization_level: OptimizationLevel::Release,
-                    debug_info: false,
-                    target_cpu: "native".to_string(),
-                    features_enabled: Vec::new(),
-                },
-            },
-            performance_metrics: metrics,
-            statistical_properties: StatisticalProperties {
-                sample_size: 100,
-                confidence_interval_95: ConfidenceInterval {
-                    lower_bound: 4000.0,
-                    upper_bound: 4400.0,
-                    confidence_level: 0.95,
-                },
-                statistical_significance: true,
-                normality_test_passed: true,
-            },
-            validation_status: BaselineValidationStatus::Valid,
-        };
-
-        let baseline_id = repo.store_baseline(baseline).unwrap();
-        let loaded_baseline = repo.load_baseline(&baseline_id).unwrap();
-        assert_eq!(loaded_baseline.baseline_id, "test_baseline");
-    }
-
-    #[test]
-    fn test_statistical_analyzer() {
-        let analyzer = StatisticalRegressionAnalyzer::new();
-
-        let baseline_metrics = PerformanceMetrics {
-            display_throughput: ThroughputMetrics {
-                mean: 4200.0,
-                median: 4150.0,
-                percentile_95: 4800.0,
-                percentile_99: 5200.0,
-                std_deviation: 250.0,
-                min: 3800.0,
-                max: 5500.0,
-            },
-            comp3_throughput: ThroughputMetrics {
-                mean: 580.0,
-                median: 575.0,
-                percentile_95: 620.0,
-                percentile_99: 650.0,
-                std_deviation: 25.0,
-                min: 540.0,
-                max: 680.0,
-            },
-            memory_usage: MemoryUsageMetrics {
-                peak_memory_mb: 245.0,
-                average_memory_mb: 180.0,
-                steady_state_memory_mb: 165.0,
-                memory_variance: 12.5,
-            },
-            latency_metrics: LatencyMetrics {
-                p50_ms: 0.24,
-                p95_ms: 0.45,
-                p99_ms: 0.68,
-                p999_ms: 1.2,
-                max_latency_ms: 2.1,
-            },
-        };
-
-        let current_metrics = PerformanceMetrics {
-            display_throughput: ThroughputMetrics {
-                mean: 4100.0, // Slightly lower
-                median: 4050.0,
-                percentile_95: 4700.0,
-                percentile_99: 5100.0,
-                std_deviation: 250.0,
-                min: 3700.0,
-                max: 5400.0,
-            },
-            comp3_throughput: ThroughputMetrics {
-                mean: 575.0, // Slightly lower
-                median: 570.0,
-                percentile_95: 615.0,
-                percentile_99: 645.0,
-                std_deviation: 25.0,
-                min: 535.0,
-                max: 675.0,
-            },
-            memory_usage: MemoryUsageMetrics {
-                peak_memory_mb: 250.0, // Slightly higher
-                average_memory_mb: 185.0,
-                steady_state_memory_mb: 170.0,
-                memory_variance: 13.0,
-            },
-            latency_metrics: LatencyMetrics {
-                p50_ms: 0.25,
-                p95_ms: 0.47,
-                p99_ms: 0.70,
-                p999_ms: 1.25,
-                max_latency_ms: 2.2,
-            },
-        };
-
-        let comparison = analyzer
-            .compare_metrics(&baseline_metrics, &current_metrics)
-            .unwrap();
-        assert!(!comparison.throughput_changes.is_empty());
-        assert!(!comparison.memory_changes.is_empty());
-        assert!(!comparison.latency_changes.is_empty());
-
-        let statistical_tests = analyzer
-            .run_statistical_tests(&baseline_metrics, &current_metrics)
-            .unwrap();
-        assert!(statistical_tests.t_test_result.degrees_of_freedom > 0);
-    }
-
-    fn statistical_test_results(significant: bool) -> StatisticalTestResults {
-        StatisticalTestResults {
-            t_test_result: TTestResult {
-                statistic: if significant { 4.0 } else { 0.5 },
-                p_value: if significant { 0.01 } else { 0.50 },
-                degrees_of_freedom: 98,
-                significant,
-            },
-            mann_whitney_u_result: MannWhitneyResult {
-                u_statistic: 1200.0,
-                p_value: if significant { 0.03 } else { 0.50 },
-                significant,
-            },
-            effect_size: EffectSize {
-                cohens_d: if significant { 0.8 } else { 0.1 },
-                glass_delta: if significant { 0.8 } else { 0.1 },
-                hedges_g: if significant { 0.79 } else { 0.1 },
-                interpretation: if significant {
-                    EffectSizeInterpretation::Large
-                } else {
-                    EffectSizeInterpretation::Negligible
-                },
-            },
-            power_analysis: PowerAnalysisResult {
-                power: 0.9,
-                required_sample_size: 64,
-                adequate_power: true,
-            },
-        }
-    }
-
-    #[test]
-    fn test_determine_regression_status_detects_improvement() {
-        let detector = PerformanceRegressionDetector::new();
-        let comparison = MetricsComparison {
-            throughput_changes: vec![MetricChange {
-                metric_name: "display_throughput_mean".to_string(),
-                baseline_value: 100.0,
-                current_value: 110.0,
-                change_percent: 10.0,
-                change_direction: ChangeDirection::Improvement,
-                statistical_significance: true,
-            }],
-            memory_changes: vec![MetricChange {
-                metric_name: "peak_memory_mb".to_string(),
-                baseline_value: 100.0,
-                current_value: 98.5,
-                change_percent: 1.5,
-                change_direction: ChangeDirection::Neutral,
-                statistical_significance: false,
-            }],
-            latency_changes: vec![MetricChange {
-                metric_name: "p95_latency_ms".to_string(),
-                baseline_value: 10.0,
-                current_value: 9.8,
-                change_percent: 2.0,
-                change_direction: ChangeDirection::Neutral,
-                statistical_significance: false,
-            }],
-            overall_change_percent: 4.5,
-        };
-        let tests = statistical_test_results(true);
-
-        let status = detector
-            .determine_regression_status(&comparison, &tests)
-            .unwrap();
-        match status {
-            RegressionStatus::Improvement { magnitude } => {
-                assert_eq!(magnitude, 10.0);
-            }
-            _ => panic!("Expected improvement status"),
-        }
-    }
-
-    #[test]
-    fn test_determine_regression_status_ignores_non_significant_improvement() {
-        let detector = PerformanceRegressionDetector::new();
-        let comparison = MetricsComparison {
-            throughput_changes: vec![MetricChange {
-                metric_name: "display_throughput_mean".to_string(),
-                baseline_value: 100.0,
-                current_value: 110.0,
-                change_percent: 10.0,
-                change_direction: ChangeDirection::Improvement,
-                statistical_significance: false,
-            }],
-            memory_changes: vec![],
-            latency_changes: vec![],
-            overall_change_percent: 10.0,
-        };
-        let tests = statistical_test_results(false);
-
-        let status = detector
-            .determine_regression_status(&comparison, &tests)
-            .unwrap();
-
-        assert!(matches!(status, RegressionStatus::NoRegression));
-    }
-
-    #[test]
-    fn test_determine_regression_status_prefers_regression_when_present() {
-        let detector = PerformanceRegressionDetector::new();
-        let comparison = MetricsComparison {
-            throughput_changes: vec![MetricChange {
-                metric_name: "display_throughput_mean".to_string(),
-                baseline_value: 100.0,
-                current_value: 95.0,
-                change_percent: 5.0,
-                change_direction: ChangeDirection::Degradation,
-                statistical_significance: true,
-            }],
-            memory_changes: vec![MetricChange {
-                metric_name: "peak_memory_mb".to_string(),
-                baseline_value: 100.0,
-                current_value: 90.0,
-                change_percent: 10.0,
-                change_direction: ChangeDirection::Improvement,
-                statistical_significance: true,
-            }],
-            latency_changes: vec![],
-            overall_change_percent: 7.5,
-        };
-        let tests = statistical_test_results(true);
-
-        let status = detector
-            .determine_regression_status(&comparison, &tests)
-            .unwrap();
-        assert!(matches!(status, RegressionStatus::MajorRegression { .. }));
-    }
-
-    #[test]
-    fn test_ci_integrator() {
-        let ci = CiIntegrator::new();
-        assert_eq!(ci.performance_gates.len(), 2);
-        assert!(
-            ci.performance_gates
-                .iter()
-                .any(|g| g.gate_id == "display_throughput")
-        );
-        assert!(
-            ci.performance_gates
-                .iter()
-                .any(|g| g.gate_id == "comp3_throughput")
-        );
-    }
-
-    #[test]
-    fn test_alert_system() {
-        let alert_system = AlertSystem::new();
-        assert_eq!(alert_system.alert_policies.len(), 1);
-        assert_eq!(alert_system.notification_channels.len(), 1);
-    }
-
-    #[test]
-    fn test_utils_create_detectors() {
-        let _standard = utils::create_standard_detector();
-        let ci = utils::create_ci_detector();
-        let dev = utils::create_dev_detector();
-
-        assert!(
-            ci.statistical_analyzer.variance_tolerance
-                < dev.statistical_analyzer.variance_tolerance
-        );
-    }
-
-    #[test]
-    fn test_generate_baseline_id_not_empty() {
-        let detector = PerformanceRegressionDetector::new();
-        let env = EnvironmentInfo {
-            rust_version: "1.92.0".to_string(),
-            target_triple: "x86_64-unknown-linux-gnu".to_string(),
-            cpu_info: CpuInfo {
-                model: "Test CPU".to_string(),
-                cores: 4,
-                frequency_mhz: 2400,
-                cache_size_kb: 8192,
-            },
-            memory_info: MemoryInfo {
-                total_gb: 16.0,
-                available_gb: 12.0,
-                memory_type: "DDR4".to_string(),
-            },
-            build_configuration: BuildConfiguration {
-                optimization_level: OptimizationLevel::Release,
-                debug_info: false,
-                target_cpu: "x86_64".to_string(),
-                features_enabled: vec!["sse2".to_string(), "avx".to_string()],
-            },
-        };
-
-        let baseline_id = detector.generate_baseline_id(&env);
-
-        // Kill mutants that return empty string or fixed string
-        assert!(!baseline_id.is_empty(), "Baseline ID must not be empty");
-        assert_ne!(baseline_id, "xyzzy", "Baseline ID must not be fixed string");
-        assert!(
-            baseline_id.len() > 10,
-            "Baseline ID should be substantial hash"
-        );
-
-        // Verify ID has expected format (kill mutants that return invalid formats)
-        assert!(
-            baseline_id.starts_with("baseline_"),
-            "Baseline ID should start with 'baseline_'"
-        );
-        assert!(
-            baseline_id.len() == 16,
-            "Baseline ID should be exactly 16 characters (8 + 8 hex chars)"
-        );
-    }
-
-    #[test]
-    fn test_establish_baseline_returns_valid_id() {
-        let mut detector = PerformanceRegressionDetector::new();
-        let metrics = create_test_performance_metrics();
-        let env = EnvironmentInfo {
-            rust_version: "1.92.0".to_string(),
-            target_triple: "x86_64-unknown-linux-gnu".to_string(),
-            cpu_info: CpuInfo {
-                model: "Test CPU".to_string(),
-                cores: 4,
-                frequency_mhz: 2400,
-                cache_size_kb: 8192,
-            },
-            memory_info: MemoryInfo {
-                total_gb: 16.0,
-                available_gb: 12.0,
-                memory_type: "DDR4".to_string(),
-            },
-            build_configuration: BuildConfiguration {
-                optimization_level: OptimizationLevel::Release,
-                debug_info: false,
-                target_cpu: "x86_64".to_string(),
-                features_enabled: vec!["sse2".to_string(), "avx".to_string()],
-            },
-        };
-
-        let result = detector.establish_baseline(metrics, env);
-        assert!(result.is_ok(), "Establish baseline should succeed");
-
-        let baseline_id = result.unwrap();
-        // Kill mutants that return empty string
-        assert!(
-            !baseline_id.is_empty(),
-            "Baseline ID from establish_baseline must not be empty"
-        );
-        assert_ne!(baseline_id, "xyzzy", "Baseline ID must not be fixed string");
-        assert!(
-            baseline_id.len() > 10,
-            "Established baseline ID should be substantial"
-        );
-    }
-
-    #[test]
-    fn test_performance_regression_detector_baseline_workflow() {
-        // Test the complete workflow to catch more mutants
-        let mut detector = PerformanceRegressionDetector::new();
-
-        let metrics = create_test_performance_metrics();
-        let env = EnvironmentInfo {
-            rust_version: "1.92.0".to_string(),
-            target_triple: "x86_64-unknown-linux-gnu".to_string(),
-            cpu_info: CpuInfo {
-                model: "Test CPU".to_string(),
-                cores: 4,
-                frequency_mhz: 2400,
-                cache_size_kb: 8192,
-            },
-            memory_info: MemoryInfo {
-                total_gb: 16.0,
-                available_gb: 12.0,
-                memory_type: "DDR4".to_string(),
-            },
-            build_configuration: BuildConfiguration {
-                optimization_level: OptimizationLevel::Release,
-                debug_info: false,
-                target_cpu: "x86_64".to_string(),
-                features_enabled: vec!["sse2".to_string(), "avx".to_string()],
-            },
-        };
-
-        // Test that baseline establishment produces a valid ID
-        let baseline_id = detector.establish_baseline(metrics, env).unwrap();
-
-        // Kill mutants that return empty or fixed strings
-        assert!(
-            !baseline_id.is_empty(),
-            "Baseline workflow should produce non-empty ID"
-        );
-        assert_ne!(
-            baseline_id, "xyzzy",
-            "Baseline workflow should not produce fixed string"
-        );
-        assert!(
-            baseline_id.len() > 10,
-            "Baseline workflow should produce substantial ID"
-        );
-
-        // Test that the detector now has some baseline data stored
-        // Just verify the workflow completed without error - the baseline_id is proof it worked
-        assert!(
-            baseline_id.starts_with("baseline_"),
-            "Baseline ID should have expected format"
-        );
-    }
-
-    fn create_test_performance_metrics() -> PerformanceMetrics {
-        PerformanceMetrics {
-            display_throughput: ThroughputMetrics {
-                mean: 4200.0,
-                median: 4100.0,
-                percentile_95: 4800.0,
-                percentile_99: 5200.0,
-                std_deviation: 300.0,
-                min: 3800.0,
-                max: 5600.0,
-            },
-            comp3_throughput: ThroughputMetrics {
-                mean: 580.0,
-                median: 575.0,
-                percentile_95: 620.0,
-                percentile_99: 650.0,
-                std_deviation: 30.0,
-                min: 540.0,
-                max: 680.0,
-            },
-            memory_usage: MemoryUsageMetrics {
-                peak_memory_mb: 245.0,
-                average_memory_mb: 180.0,
-                steady_state_memory_mb: 165.0,
-                memory_variance: 12.5,
-            },
-            latency_metrics: LatencyMetrics {
-                p50_ms: 0.24,
-                p95_ms: 0.45,
-                p99_ms: 0.68,
-                p999_ms: 1.2,
-                max_latency_ms: 2.1,
-            },
-        }
-    }
-}
+mod tests;
